successfully implemented state management with useReducer and context api
