+++ src/components/Progress.js
@@ -1,4 +1,7 @@
-function Progress({ index, numQuestions, points, maxPossiblePoints, answer }) {
+import { useQuiz } from "../context/QuizContextProcider";
+
+function Progress() {
+  const { index, numQuestions, points, maxPossiblePoints, answer } = useQuiz();
   return (
     <header className="progress">
       <progress
