+++ src/components/Question.js
@@ -1,6 +1,9 @@
+import { useQuiz } from "../context/QuizContextProcider";
 import Options from "./Options";
 
-export default function Question({ question, dispatch, answer }) {
+export default function Question() {
+  const { questions, index, dispatch, answer } = useQuiz();
+  const question = questions[index];
   return (
     <div>
       <h4>{question.question}</h4>
