+++ src/components/StartScreen.js
@@ -1,4 +1,7 @@
-export default function StartScreen({ numQuestions, dispatch }) {
+import { useQuiz } from "../context/QuizContextProcider";
+
+export default function StartScreen() {
+  const { numQuestions, dispatch } = useQuiz();
   return (
     <div className="start">
       <h2>Welcome to the React Quiz</h2>
